+++ pages/login/index.js
@@ -1,43 +1,50 @@
 import React, { useState } from "react";
-import Layout from "../Components/Layout";
-import { Typography, TextField, Button, Box } from "@mui/material";
+import Layout from "../../Components/Layout";
+import { Typography, TextField,Box } from "@mui/material";
 import IconButton from "@mui/material/IconButton";
 import OutlinedInput from "@mui/material/OutlinedInput";
 import InputLabel from "@mui/material/InputLabel";
 import InputAdornment from "@mui/material/InputAdornment";
 import Visibility from "@mui/icons-material/Visibility";
 import VisibilityOff from "@mui/icons-material/VisibilityOff";
-import { LoadingButton } from "@mui/lab";
+import clienteAxios from "../../utils/axios";
 import SendIcon from "@mui/icons-material/Send";
+import { LoadingButton } from "@mui/lab";
 import { useRouter } from "next/router";
-import { useForm } from "../hooks/useForm";
-import clienteAxios from "../utils/axios";
+import { useForm } from "../../hooks/useForm";
+import  {AuthContext, } from "../../Context/AuthContext";
+import { useContext } from "react";
+import {useLocalStorage} from "../../hooks/useLocalStorage"
 
-const Admin = () => {
+
+const Login = () => {
   const router = useRouter();
   const [loading, setLoading] = useState(false);
   const [values, setValues] = React.useState({
     password: "",
     showPassword: false,
   });
+
   const initialForm = {
-    register: true,
-    name: "",
-    lastName: "",
     email: "",
     password: "",
   };
-  const [user, actualizarState, reset] = useForm(initialForm);
+  const [login, actualizarState, reset] = useForm(initialForm);
+  const [auth,guardarAuth] = useContext(AuthContext);
+  const [valToken,setToken] = useLocalStorage('userVal');
+
 
   const handleChange = (prop) => (event) => {
     setValues({ ...values, [prop]: event.target.value });
   };
+
   const handleClickShowPassword = () => {
     setValues({
       ...values,
       showPassword: !values.showPassword,
     });
   };
+
   const handleMouseDownPassword = (event) => {
     event.preventDefault();
   };
@@ -46,12 +53,15 @@ const Admin = () => {
     e.preventDefault();
     setLoading(true);
     clienteAxios
-      .post("/admin", user)
+      .post("/login", login)
       .then((response) => {
         //console.log(response.data);
-        const { status } = response.data;
-        if (status) {
-          router.push("/login"); //dirigir a la pagina de inicio
+        const { auth,token,infoUser } = response.data;
+        if (auth) {
+          guardarAuth({token,auth,infoUser});
+          setToken({token,auth,infoUser});
+          //console.log(token)
+          router.push("/agenda"); //dirigir a la pagina de inicio
           //document.querySelector("#form").reset();
         }
       })
@@ -78,37 +88,9 @@ const Admin = () => {
           }}
         >
           <Typography component="h6" variant="h6">
-            Crea Cuenta{" "}
+            Inicia Sesion{" "}
           </Typography>
           <form id="form" onSubmit={handlerSubmit}>
-            <div >
-              <TextField
-                 sx={{ m: 1, width: '25ch' }}
-                //  variant="outlined"
-                // fullWidth
-                size="small"
-                required
-                id="name"
-                label="Name"
-                name="name"
-                inputProps={{ type: "text" }}
-                onChange={actualizarState}
-              ></TextField>
-            </div>
-            <div >
-              <TextField
-                sx={{ m: 1, width: '25ch' }}
-                //  variant="outlined"
-                //fullWidth
-                size="small"
-                required
-                id="lastName"
-                label="last name"
-                name="lastName"
-                inputProps={{ type: "text" }}
-                onChange={actualizarState}
-              ></TextField>
-            </div>
             <div>
               <TextField
                 required
@@ -169,7 +151,7 @@ const Admin = () => {
                 variant="contained"
                 type="submit"
               >
-                Crear Cuenta
+                Enviar
               </LoadingButton>
             </div>
           </form>
@@ -179,4 +161,4 @@ const Admin = () => {
   );
 };
 
-export default Admin;
+export default Login;
